requirements.txt
pydantic_ai
python-dotenv
openai
streamlit
